backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Document processing
python-docx==1.1.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
Pillow==10.2.0
pytesseract==0.3.10

# Text processing
spacy==3.7.2
python-dotenv==1.0.0

# Supabase
supabase==2.3.0
httpx==0.26.0
